requirements.txt
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.13.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pkg-resources==0.0.0
PyJWT==1.4.2
python-dotenv==0.9.1
python-magic==0.4.15
Werkzeug==0.14.1
